Refactored go code. Is now more idiomatic, easier to read, and logically placed. Need to refactore front-end JS to match with changes to urls and the json keys changing via tags in go code. This addresses issue #15 and #18
