+++ application.properties.example
@@ -3,5 +3,6 @@ dbpassword=mongotest
 dbport=0000
 dbname=test
 dbaddress=0.0.0.0
+userCollection=users
 loglevel=info
 port=8080
